+++ src/cluster/flat.rs
@@ -491,6 +491,7 @@ mod tests {
 
         // DenStream
         let mut ds = DenStream::new(2.0, 2)
+            .with_initial_buffer_size(0)
             .with_beta(0.5)
             .with_lambda(0.001)
             .with_mu(3.0);